+++ CHANGELOG.md
@@ -31,8 +31,11 @@ if you don't want to use any ``DatasetVersion``.
 * When using ``ConfigurationType.supplement`` default values are now ignored when determining if values exist.
 * The ``pipeline`` attributes gets replaced by the ``task_results`` attributes for ``Document``, ``Collection`` and
   ``DatasetVersion``.
-* Default property to write contributions to has become equal to the ``growth_phase`` and
-  will be set in the ``derivatives`` field instead of ``properties``.
+* When writing contributions to ``Documents`` the default field is now ``derivatives``.
+  Furthermore a key equal to the ``growth_phase`` is automatically added to the ``derivatives`` dictionary.
+  The value for this key is always an empty dictionary. Any ``to_property`` configuration will write to this dictionary.
+  Otherwise contributions get merged into the dictionary.  It's still possible to write to ``properties`` without
+  adding special ``growth_phase`` keys for backward compatability.
 * If ``ResourceGrowthProcessor`` encounters multiple ``Resources`` per ``Document`` or
   if a single ``Resource`` yields multiple results. Then the ``reduce_contributions`` method will be called to
   determine how contribution data from ``Resources`` should compliment ``Document`` data. The default is to only use